chore: move diretório /app para dentro do /src
