+++ src/app/(pages)/projetos/estatisticas/page.tsx
@@ -1,0 +1,5 @@
+export type Props = {}
+
+export default function CompanyStatistics(): JSX.Element {
+  return <div></div>
+}
